Network live under insights namespace
