+++ lib/rspec/buildkite/insights/uploader.rb
@@ -8,7 +8,7 @@ require "websocket"
 
 require_relative "tracer"
 
-require_relative "../instrumentations/network"
+require_relative "network"
 
 require "active_support"
 require "active_support/notifications"
@@ -251,7 +251,7 @@ module RSpec::Buildkite::Insights
         end
       end
 
-      RSpec::Buildkite::Insights::Instrumentations::Network.configure
+      RSpec::Buildkite::Insights::Network.configure
 
       ActiveSupport::Notifications.subscribe("sql.active_record") do |name, start, finish, id, payload|
         tracer&.backfill(:sql, finish - start, { query: payload[:sql] })
